# specs/001-ai-ai/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: VibeMusic API - 감정 기반 AI 음악 생성
  description: 키보드 타이핑 패턴을 분석하여 감정을 캡처하고 AI 음악을 생성하는 API
  version: 1.0.0
  contact:
    name: VibeMusic Team
  license:
    name: MIT
servers:
  - url: https://api.vibemusic.app/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
paths:
...
  /sessions/{session_id}/music/{music_id}/download:
    get:
      summary: 생성된 음악 파일 다운로드
      tags:
        - Generation
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: music_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - SessionToken: []
      responses:
        '200':
          description: 음악 파일
          content:
            audio/wav:
              schema:
                type: string
                format: binary
            audio/mpeg:
              schema:
                type: string
                format: binary
            audio/flac:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    SessionToken:
      type: http
      scheme: bearer
      description: 세션 생성 시 받은 토큰 사용
...